third_party/blink/web_tests/platform/generic/fast/dom/DOMException/XPathException-expected.txt

Tests the properties of the exception thrown when using XPath.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.toString() is "TypeError: Failed to read the 'numberValue' property from 'XPathResult': The result type is not a number."
PASS Object.prototype.toString.call(e) is "[object Error]"
PASS Object.prototype.toString.call(e.__proto__) is "[object Object]"
PASS e.constructor.toString() is "function TypeError() { [native code] }"
PASS e.constructor is window.TypeError
PASS successfullyParsed is true

TEST COMPLETE
